Integrate code to title or hide references section.
This automatically adds a title to the reference section pandoc-citeproc
produces. It takes `reference-section-title` metadata as the reference section
title. If this metadata field is not specified, there is no automatic header.
So existing documents will continue to work as they always have.

It can also take a boolean metadtata `remove-ref-section` to remove the
reference section altogether. This can be useful for situations when you want
citations to be resolved, but don't want an extra few pages for references at
the end of the document.