Changes to title-short logic and test cases.
Nick Bart says:

"The logic should be: get the "shorttitle" field, unless the entry also has a
maintitle AND is some kind of book (in other words, is not isChapterlike).

(For a book with a title and a maintitle, the maintitle becomes CSL title,
and title becomes CSL volume-title, hence shorttitle cannot be mapped
anywhere (there is no CSL volume-title-short). There’s also no biblatex
mainshorttitle, but we could try to map the maintitle only if there’s also
a mainsubtitle, or the maintitle up to the first colon, to CSL title-short
in such cases; not sure whether we ever discussed this.)"