Removed compressName transform on json output.
Closes #169.

This transformation prevented the json produced from
using a fully broken-out specification of the author,
instead pushing suffixes and prefixes into the names
themselves.

I'm not sure why this transformation was originally added,
but it seems to have been needed for older versions of zotero.
It doesn't seem needed or wanted any more.