support citation-label variable (fix #160)
The `citation-label` was not rendered, even when this label was
explicitly set in the bibliography. This is now fixed. Citation labels
are modelled after citation numbers. In particular, citation labels can
also be hyperlinked.

But since this was largely done by cargo-cult programming, there may be
some bugs in this fix. In particular, I have no idea how this interacts
with collapsing.

**API changes:**

In `Text.CSL,Reference`: The exported type `CLabel` describes a label
name in the CSL variables. This allows us to distinguish labels from
ordinary strings. This type is analogous to `CNum`.

In `Text.CSL.Style`: In `Output` the `OCitLabel` case describes a label
in the rendered output. This case is analogous to `OCitNum`.