Made bibtex parsing more efficient.  (API change)
This commit adds an extra parameter to the bibtex-parsing
functions: readBibtex, readBibtexString, readBiblioString,
readBiblioFile.  The extra (first) parameter is a function
from strings to booleans that acts as a filter on citation
identifiers, so that non-matching identifiers are not parsed
in full.  Pandoc-citeproc now gets a list of the citation
identifiers in a document and passes it to readBiblio*,
which gives a very substantial speed increase with large
bibliographies.