locators: allows parsing balanced brackets in locators like 1(2)(a)(i)
This is useful for citing legislation. It is necessary to be more liberal
in locator parsing because some styles put other elements after a locator,
like short titles in the case of AGLC, and if you can't parse a locator,
it gets rendered after, or worse, split in half.

this includes a minor (single-line) bugfix for #361 where ibids were overly-disambiguated on their locators which had different whitespace depending on whether you put a citeLabel before it (ie p. 89 -> " 89" but 89 -> "89")